third_party/blink/web_tests/fast/css/pseudo-valid-004-expected.txt

This test performs a check for the :valid/:invalid CSS selector on textarea elements where it shouldn't be applied.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

 
PASS document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color') is 'rgb(0, 255, 0)'
PASS document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color') is 'rgb(0, 255, 0)'
PASS successfullyParsed is true

TEST COMPLETE
